pyproject.toml
[project]
name = "slides-xp"
version = "0.1.0"
description = ""
authors = [
    {name = "Maddy Guthridge",email = "hello@maddyguthridge.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pyhtml-enhanced (>=2.2.0,<3.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "mistune (>=3.1.1,<4.0.0)",
    "pygments (>=2.19.1,<3.0.0)"
]

[tool.poetry.scripts]
sxp = "slides_xp.__main__:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
